+++ bitset.h
@@ -3,11 +3,11 @@
 
 #include <stdint.h>
 
-uint8_t popcnt8(uint8_t x);
-void bitset_set(uint8_t *s, uint8_t idx);
-void bitset_clear(uint8_t *s, uint8_t idx);
-int bitset_test(const uint8_t *s, uint8_t idx);
-uint8_t bitset_count(const uint8_t *s, uint8_t size);
-int bitset_findfirst(const uint8_t *s, uint8_t size);
+unsigned popcnt8(uint8_t x);
+void bitset_set(uint8_t *s, unsigned idx);
+void bitset_clear(uint8_t *s, unsigned idx);
+int bitset_test(const uint8_t *s, unsigned idx);
+unsigned bitset_count(const uint8_t *s, unsigned size);
+int bitset_findfirst(const uint8_t *s, unsigned size);
 
 #endif /* BITSET_H */
